Tools/parser/unparse.py
"Usage: unparse.py <path to source file>"
zaimportuj sys
zaimportuj ast
zaimportuj tokenize
zaimportuj io
zaimportuj os

# Large float oraz imaginary literals get turned into infinities w the AST.
# We unparse those infinities to INFSTR.
INFSTR = "1e" + repr(sys.float_info.max_10_exp + 1)

def interleave(inter, f, seq):
    """Call f on each item w seq, calling inter() w between.
    """
    seq = iter(seq)
    spróbuj:
        f(next(seq))
    wyjąwszy StopIteration:
        dalej
    inaczej:
        dla x w seq:
            inter()
            f(x)

klasa Unparser:
    """Methods w this klasa recursively traverse an AST oraz
    output source code dla the abstract syntax; original formatting
    jest disregarded. """

    def __init__(self, tree, file = sys.stdout):
        """Unparser(tree, file=sys.stdout) -> Nic.
         Print the source dla tree to file."""
        self.f = file
        self._indent = 0
        self.dispatch(tree)
        print("", file=self.f)
        self.f.flush()

    def fill(self, text = ""):
        "Indent a piece of text, according to the current indentation level"
        self.f.write("\n"+"    "*self._indent + text)

    def write(self, text):
        "Append a piece of text to the current line."
        self.f.write(text)

    def enter(self):
        "Print ':', oraz increase the indentation."
        self.write(":")
        self._indent += 1

    def leave(self):
        "Decrease the indentation level."
        self._indent -= 1

    def dispatch(self, tree):
        "Dispatcher function, dispatching tree type T to method _T."
        jeżeli isinstance(tree, list):
            dla t w tree:
                self.dispatch(t)
            zwróć
        meth = getattr(self, "_"+tree.__class__.__name__)
        meth(tree)


    ############### Unparsing methods ######################
    # There should be one method per concrete grammar type #
    # Constructors should be grouped by sum type. Ideally, #
    # this would follow the order w the grammar, but      #
    # currently doesn't.                                   #
    ########################################################

    def _Module(self, tree):
        dla stmt w tree.body:
            self.dispatch(stmt)

    # stmt
    def _Expr(self, tree):
        self.fill()
        self.dispatch(tree.value)

    def _Import(self, t):
        self.fill("zaimportuj ")
        interleave(lambda: self.write(", "), self.dispatch, t.names)

    def _ImportFrom(self, t):
        self.fill("z ")
        self.write("." * t.level)
        jeżeli t.module:
            self.write(t.module)
        self.write(" zaimportuj ")
        interleave(lambda: self.write(", "), self.dispatch, t.names)

    def _Assign(self, t):
        self.fill()
        dla target w t.targets:
            self.dispatch(target)
            self.write(" = ")
        self.dispatch(t.value)

    def _AugAssign(self, t):
        self.fill()
        self.dispatch(t.target)
        self.write(" "+self.binop[t.op.__class__.__name__]+"= ")
        self.dispatch(t.value)

    def _Return(self, t):
        self.fill("zwróć")
        jeżeli t.value:
            self.write(" ")
            self.dispatch(t.value)

    def _Pass(self, t):
        self.fill("dalej")

    def _Break(self, t):
        self.fill("przerwij")

    def _Continue(self, t):
        self.fill("kontynuuj")

    def _Delete(self, t):
        self.fill("usuń ")
        interleave(lambda: self.write(", "), self.dispatch, t.targets)

    def _Assert(self, t):
        self.fill("assert ")
        self.dispatch(t.test)
        jeżeli t.msg:
            self.write(", ")
            self.dispatch(t.msg)

    def _Global(self, t):
        self.fill("global ")
        interleave(lambda: self.write(", "), self.write, t.names)

    def _Nonlocal(self, t):
        self.fill("nonlocal ")
        interleave(lambda: self.write(", "), self.write, t.names)

    def _Await(self, t):
        self.write("(")
        self.write("await")
        jeżeli t.value:
            self.write(" ")
            self.dispatch(t.value)
        self.write(")")

    def _Yield(self, t):
        self.write("(")
        self.write("uzyskaj")
        jeżeli t.value:
            self.write(" ")
            self.dispatch(t.value)
        self.write(")")

    def _YieldFrom(self, t):
        self.write("(")
        self.write("uzyskaj from")
        jeżeli t.value:
            self.write(" ")
            self.dispatch(t.value)
        self.write(")")

    def _Raise(self, t):
        self.fill("podnieś")
        jeżeli nie t.exc:
            assert nie t.cause
            zwróć
        self.write(" ")
        self.dispatch(t.exc)
        jeżeli t.cause:
            self.write(" z ")
            self.dispatch(t.cause)

    def _Try(self, t):
        self.fill("spróbuj")
        self.enter()
        self.dispatch(t.body)
        self.leave()
        dla ex w t.handlers:
            self.dispatch(ex)
        jeżeli t.orinaczej:
            self.fill("inaczej")
            self.enter()
            self.dispatch(t.orinaczej)
            self.leave()
        jeżeli t.finalbody:
            self.fill("w_końcu")
            self.enter()
            self.dispatch(t.finalbody)
            self.leave()

    def _ExceptHandler(self, t):
        self.fill("wyjąwszy")
        jeżeli t.type:
            self.write(" ")
            self.dispatch(t.type)
        jeżeli t.name:
            self.write(" jako ")
            self.write(t.name)
        self.enter()
        self.dispatch(t.body)
        self.leave()

    def _ClassDef(self, t):
        self.write("\n")
        dla deco w t.decorator_list:
            self.fill("@")
            self.dispatch(deco)
        self.fill("class "+t.name)
        self.write("(")
        comma = Nieprawda
        dla e w t.bases:
            jeżeli comma: self.write(", ")
            inaczej: comma = Prawda
            self.dispatch(e)
        dla e w t.keywords:
            jeżeli comma: self.write(", ")
            inaczej: comma = Prawda
            self.dispatch(e)
        self.write(")")

        self.enter()
        self.dispatch(t.body)
        self.leave()

    def _FunctionDef(self, t):
        self.__FunctionDef_helper(t, "def")

    def _AsyncFunctionDef(self, t):
        self.__FunctionDef_helper(t, "async def")

    def __FunctionDef_helper(self, t, fill_suffix):
        self.write("\n")
        dla deco w t.decorator_list:
            self.fill("@")
            self.dispatch(deco)
        def_str = fill_suffix+" "+t.name + "("
        self.fill(def_str)
        self.dispatch(t.args)
        self.write(")")
        jeżeli t.returns:
            self.write(" -> ")
            self.dispatch(t.returns)
        self.enter()
        self.dispatch(t.body)
        self.leave()

    def _For(self, t):
        self.__For_helper("dla ", t)

    def _AsyncFor(self, t):
        self.__For_helper("async dla ", t)

    def __For_helper(self, fill, t):
        self.fill(fill)
        self.dispatch(t.target)
        self.write(" w ")
        self.dispatch(t.iter)
        self.enter()
        self.dispatch(t.body)
        self.leave()
        jeżeli t.orinaczej:
            self.fill("inaczej")
            self.enter()
            self.dispatch(t.orinaczej)
            self.leave()

    def _If(self, t):
        self.fill("jeżeli ")
        self.dispatch(t.test)
        self.enter()
        self.dispatch(t.body)
        self.leave()
        # collapse nested ifs into equivalent elifs.
        dopóki (t.orinaczej oraz len(t.orinaczej) == 1 oraz
               isinstance(t.orinaczej[0], ast.If)):
            t = t.orinaczej[0]
            self.fill("albo_inaczej ")
            self.dispatch(t.test)
            self.enter()
            self.dispatch(t.body)
            self.leave()
        # final inaczej
        jeżeli t.orinaczej:
            self.fill("inaczej")
            self.enter()
            self.dispatch(t.orinaczej)
            self.leave()

    def _While(self, t):
        self.fill("dopóki ")
        self.dispatch(t.test)
        self.enter()
        self.dispatch(t.body)
        self.leave()
        jeżeli t.orinaczej:
            self.fill("inaczej")
            self.enter()
            self.dispatch(t.orinaczej)
            self.leave()

    def _With(self, t):
        self.fill("przy ")
        interleave(lambda: self.write(", "), self.dispatch, t.items)
        self.enter()
        self.dispatch(t.body)
        self.leave()

    def _AsyncWith(self, t):
        self.fill("async przy ")
        interleave(lambda: self.write(", "), self.dispatch, t.items)
        self.enter()
        self.dispatch(t.body)
        self.leave()

    # expr
    def _Bytes(self, t):
        self.write(repr(t.s))

    def _Str(self, tree):
        self.write(repr(tree.s))

    def _Name(self, t):
        self.write(t.id)

    def _NameConstant(self, t):
        self.write(repr(t.value))

    def _Num(self, t):
        # Substitute overflowing decimal literal dla AST infinities.
        self.write(repr(t.n).replace("inf", INFSTR))

    def _List(self, t):
        self.write("[")
        interleave(lambda: self.write(", "), self.dispatch, t.elts)
        self.write("]")

    def _ListComp(self, t):
        self.write("[")
        self.dispatch(t.elt)
        dla gen w t.generators:
            self.dispatch(gen)
        self.write("]")

    def _GeneratorExp(self, t):
        self.write("(")
        self.dispatch(t.elt)
        dla gen w t.generators:
            self.dispatch(gen)
        self.write(")")

    def _SetComp(self, t):
        self.write("{")
        self.dispatch(t.elt)
        dla gen w t.generators:
            self.dispatch(gen)
        self.write("}")

    def _DictComp(self, t):
        self.write("{")
        self.dispatch(t.key)
        self.write(": ")
        self.dispatch(t.value)
        dla gen w t.generators:
            self.dispatch(gen)
        self.write("}")

    def _comprehension(self, t):
        self.write(" dla ")
        self.dispatch(t.target)
        self.write(" w ")
        self.dispatch(t.iter)
        dla if_clause w t.ifs:
            self.write(" jeżeli ")
            self.dispatch(if_clause)

    def _IfExp(self, t):
        self.write("(")
        self.dispatch(t.body)
        self.write(" jeżeli ")
        self.dispatch(t.test)
        self.write(" inaczej ")
        self.dispatch(t.orinaczej)
        self.write(")")

    def _Set(self, t):
        assert(t.elts) # should be at least one element
        self.write("{")
        interleave(lambda: self.write(", "), self.dispatch, t.elts)
        self.write("}")

    def _Dict(self, t):
        self.write("{")
        def write_pair(pair):
            (k, v) = pair
            self.dispatch(k)
            self.write(": ")
            self.dispatch(v)
        interleave(lambda: self.write(", "), write_pair, zip(t.keys, t.values))
        self.write("}")

    def _Tuple(self, t):
        self.write("(")
        jeżeli len(t.elts) == 1:
            (elt,) = t.elts
            self.dispatch(elt)
            self.write(",")
        inaczej:
            interleave(lambda: self.write(", "), self.dispatch, t.elts)
        self.write(")")

    unop = {"Invert":"~", "Not": "not", "UAdd":"+", "USub":"-"}
    def _UnaryOp(self, t):
        self.write("(")
        self.write(self.unop[t.op.__class__.__name__])
        self.write(" ")
        self.dispatch(t.operand)
        self.write(")")

    binop = { "Add":"+", "Sub":"-", "Mult":"*", "MatMult":"@", "Div":"/", "Mod":"%",
                    "LShift":"<<", "RShift":">>", "BitOr":"|", "BitXor":"^", "BitAnd":"&",
                    "FloorDiv":"//", "Pow": "**"}
    def _BinOp(self, t):
        self.write("(")
        self.dispatch(t.left)
        self.write(" " + self.binop[t.op.__class__.__name__] + " ")
        self.dispatch(t.right)
        self.write(")")

    cmpops = {"Eq":"==", "NotEq":"!=", "Lt":"<", "LtE":"<=", "Gt":">", "GtE":">=",
                        "Is":"is", "IsNot":"is not", "In":"in", "NotIn":"not in"}
    def _Compare(self, t):
        self.write("(")
        self.dispatch(t.left)
        dla o, e w zip(t.ops, t.comparators):
            self.write(" " + self.cmpops[o.__class__.__name__] + " ")
            self.dispatch(e)
        self.write(")")

    boolops = {ast.And: 'and', ast.Or: 'or'}
    def _BoolOp(self, t):
        self.write("(")
        s = " %s " % self.boolops[t.op.__class__]
        interleave(lambda: self.write(s), self.dispatch, t.values)
        self.write(")")

    def _Attribute(self,t):
        self.dispatch(t.value)
        # Special case: 3.__abs__() jest a syntax error, so jeżeli t.value
        # jest an integer literal then we need to either parenthesize
        # it albo add an extra space to get 3 .__abs__().
        jeżeli isinstance(t.value, ast.Num) oraz isinstance(t.value.n, int):
            self.write(" ")
        self.write(".")
        self.write(t.attr)

    def _Call(self, t):
        self.dispatch(t.func)
        self.write("(")
        comma = Nieprawda
        dla e w t.args:
            jeżeli comma: self.write(", ")
            inaczej: comma = Prawda
            self.dispatch(e)
        dla e w t.keywords:
            jeżeli comma: self.write(", ")
            inaczej: comma = Prawda
            self.dispatch(e)
        self.write(")")

    def _Subscript(self, t):
        self.dispatch(t.value)
        self.write("[")
        self.dispatch(t.slice)
        self.write("]")

    def _Starred(self, t):
        self.write("*")
        self.dispatch(t.value)

    # slice
    def _Ellipsis(self, t):
        self.write("...")

    def _Index(self, t):
        self.dispatch(t.value)

    def _Slice(self, t):
        jeżeli t.lower:
            self.dispatch(t.lower)
        self.write(":")
        jeżeli t.upper:
            self.dispatch(t.upper)
        jeżeli t.step:
            self.write(":")
            self.dispatch(t.step)

    def _ExtSlice(self, t):
        interleave(lambda: self.write(', '), self.dispatch, t.dims)

    # argument
    def _arg(self, t):
        self.write(t.arg)
        jeżeli t.annotation:
            self.write(": ")
            self.dispatch(t.annotation)

    # others
    def _arguments(self, t):
        first = Prawda
        # normal arguments
        defaults = [Nic] * (len(t.args) - len(t.defaults)) + t.defaults
        dla a, d w zip(t.args, defaults):
            jeżeli first:first = Nieprawda
            inaczej: self.write(", ")
            self.dispatch(a)
            jeżeli d:
                self.write("=")
                self.dispatch(d)

        # varargs, albo bare '*' jeżeli no varargs but keyword-only arguments present
        jeżeli t.vararg albo t.kwonlyargs:
            jeżeli first:first = Nieprawda
            inaczej: self.write(", ")
            self.write("*")
            jeżeli t.vararg:
                self.write(t.vararg.arg)
                jeżeli t.vararg.annotation:
                    self.write(": ")
                    self.dispatch(t.vararg.annotation)

        # keyword-only arguments
        jeżeli t.kwonlyargs:
            dla a, d w zip(t.kwonlyargs, t.kw_defaults):
                jeżeli first:first = Nieprawda
                inaczej: self.write(", ")
                self.dispatch(a),
                jeżeli d:
                    self.write("=")
                    self.dispatch(d)

        # kwargs
        jeżeli t.kwarg:
            jeżeli first:first = Nieprawda
            inaczej: self.write(", ")
            self.write("**"+t.kwarg.arg)
            jeżeli t.kwarg.annotation:
                self.write(": ")
                self.dispatch(t.kwarg.annotation)

    def _keyword(self, t):
        jeżeli t.arg jest Nic:
            self.write("**")
        inaczej:
            self.write(t.arg)
            self.write("=")
        self.dispatch(t.value)

    def _Lambda(self, t):
        self.write("(")
        self.write("lambda ")
        self.dispatch(t.args)
        self.write(": ")
        self.dispatch(t.body)
        self.write(")")

    def _alias(self, t):
        self.write(t.name)
        jeżeli t.asname:
            self.write(" jako "+t.asname)

    def _withitem(self, t):
        self.dispatch(t.context_expr)
        jeżeli t.optional_vars:
            self.write(" jako ")
            self.dispatch(t.optional_vars)

def roundtrip(filename, output=sys.stdout):
    przy open(filename, "rb") jako pyfile:
        encoding = tokenize.detect_encoding(pyfile.readline)[0]
    przy open(filename, "r", encoding=encoding) jako pyfile:
        source = pyfile.read()
    tree = compile(source, filename, "exec", ast.PyCF_ONLY_AST)
    Unparser(tree, output)



def testdir(a):
    spróbuj:
        names = [n dla n w os.listdir(a) jeżeli n.endswith('.py')]
    wyjąwszy OSError:
        print("Directory nie readable: %s" % a, file=sys.stderr)
    inaczej:
        dla n w names:
            fullname = os.path.join(a, n)
            jeżeli os.path.isfile(fullname):
                output = io.StringIO()
                print('Testing %s' % fullname)
                spróbuj:
                    roundtrip(fullname, output)
                wyjąwszy Exception jako e:
                    print('  Failed to compile, exception jest %s' % repr(e))
            albo_inaczej os.path.isdir(fullname):
                testdir(fullname)

def main(args):
    jeżeli args[0] == '--testdir':
        dla a w args[1:]:
            testdir(a)
    inaczej:
        dla a w args:
            roundtrip(a)

jeżeli __name__=='__main__':
    main(sys.argv[1:])
